Reproject geometries in layers with non-project CRS
Reprojection is now done explicitly on the QgsRubberBand with flags and QgsCoordinateTransforms. Also the connection between the signal rbFinished and slot draw() has been revised so that the geometry emitted by rbFinished is correctly passed to draw(). Previously, the geometry was not being passed, and draw() was taking the geometry directly from QgsRubberBand, and QgsRubberBand's geometry was not being correctly transformed. This bug was really hard to solve!