Remove automatic memory layer creation
This is no longer necessary because tool activation and use is confined to situations when Polygon vector layers are selected and editing is turned on. In the future, it may be a good idea to add automatic layer creation back in, and trigger it when the user has selected no layers or a raster layer.