Fix bug in rollback after merge
Rollback now seems to produce correct geometries in all tested cases, but the user still has to hit undo multiple times to undo multiple holes or a merge. From a user's perspective, this will look like unexpected behavior, so we should document it. Ideally, we would modify the default behavior of QgsVectorLayerEditBuffer to group some of these changes together, so that undo would apply to all of them at once. I may take that approach in a later commit.